Arm64 : Adjust Fortran nested Hack to use bit 1 on aarch64, add to C
This addresses Issue #32 by adjusting the nested function hack to use
bit 1 rather than bit 2.

NOTE: that Arm reserve both bits 0 and 1 for furture use, so that this
is no less a hack than using bit 2 - but it might be a somewhat more
compatible hack.

Tested with the example in Issue #32 (otherwise, there's no real change
in the testsuite).